stt-service/stt_loaders/openai_realtime_loader.py
#!/usr/bin/env python3
"""
OpenAI Realtime API model loader for VoxScribe (Example)
"""

from typing import Any, Tuple
from pathlib import Path
from .base_loader import BaseModelLoader


class OpenAIRealtimeLoader(BaseModelLoader):
    """Loader for OpenAI Realtime API models."""
    
    def __init__(self):
        super().__init__("openai_realtime")
    
    def check_dependencies(self) -> Tuple[bool, str]:
        """Check if OpenAI dependencies are available."""
        try:
            import openai
            import os
            
            # Check for API key
            if not os.getenv("OPENAI_API_KEY"):
                return False, "OPENAI_API_KEY environment variable not set"
            
            return True, ""
        except ImportError as e:
            return False, f"OpenAI dependencies not available: {e}"
    
    def load_model(self, model_id: str, cache_dir: Path, device: str) -> Any:
        """Load OpenAI Realtime API client."""
        import openai
        
        # For API-based models, we just return a configured client
        client = openai.OpenAI()
        
        # Return a wrapper object with the model configuration
        return {
            "client": client,
            "model_id": model_id,
            "type": "api_based"
        }
    
    def get_model_info(self, model_id: str) -> dict:
        """Get information about an OpenAI Realtime model."""
        model_info_map = {
            "gpt-4o-realtime-preview": {
                "display_name": "GPT-4o Realtime Preview",
                "size": "API-based",
                "description": "OpenAI GPT-4o Realtime API for speech"
            },
            "gpt-4o-realtime": {
                "display_name": "GPT-4o Realtime",
                "size": "API-based", 
                "description": "OpenAI GPT-4o Realtime API for speech"
            }
        }
        
        info = model_info_map.get(model_id, {
            "display_name": model_id,
            "size": "API-based",
            "description": f"OpenAI Realtime model {model_id}"
        })
        
        info.update({
            "supports_timestamps": True,
            "supports_word_timestamps": False,
            "is_api_based": True,
            "requires_internet": True
        })
        
        return info