stt-service/requirements.txt
# STT Service Requirements
# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
torch>=2.0.0
torchaudio>=2.9.0
soundfile>=0.13.1
numpy>=1.20.0
pandas>=1.3.0
pydub>=0.25.1
ffmpeg-python>=0.2.0
packaging>=21.0
httpx>=0.24.0

# Whisper dependencies
openai-whisper>=20250625

# NeMo dependencies (Parakeet and Canary-Qwen)
nemo_toolkit[asr] @ git+https://github.com/NVIDIA/NeMo.git
megatron-core
sacrebleu
cuda-python>=12.3
sentencepiece>=0.2.0
mistral-common
fasttext-wheel
